+++ package.json
@@ -12,6 +12,7 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.6.2",
+    "draft-convert": "^2.1.13",
     "fullcalendar": "^6.1.10",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
